Move complexity of setting up stages into FeistelCipherBuilder
In essence the feistel cipher just needs a list of stages.
How those stages are created is not a concern of the cipher.
The newly created builder is just one way to create the stages.
Users of the cipher are not restricted to just that builder.
(That's why the builder is outside of the FeistelCipher class.)